requirements.txt
tracemoepy
sqlalchemy==1.3.23
pymongo 
aiohttp
aiofiles
asyncio
dnspython
ffmpeg-python
future
git+https://github.com/kaalmous/downloader@kaal
motor
pillow
psutil
pyrogram==1.4.16
py-tgcalls==0.8.6
python-dotenv
requests
tgcrypto
wget
youtube-search
youtube-search-python
pymediainfo

